Minor revisions fixing TMDN training.
* MDN: changed a safe-minimizing parameter, based on empirical results.
* TDMN: delta weights' placeholder was turned to float64.
* MLPG: got rid of potential cases of log(0) and division by 0.

+++ neural_networks/models/_tmdn.py
@@ -52,7 +52,7 @@ class TrajectoryMDN(MixtureDensityNetwork):
         """Build the instance's placeholders."""
         super()._build_placeholders()
         self._holders['_delta_weights'] = (
-            tf.placeholder(tf.float32, [None, None])
+            tf.placeholder(tf.float64, [None, None])
         )
 
     @onetimemethod
